pyproject.toml
[project]
name = "traffic-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "django-cors-headers>=4.3.1",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "mysqlclient>=2.2.4",
    "numpy>=2.3.0",
    "ipykernel>=6.29.5",
    "openpyxl>=3.1.5",
    "fastkml>=1.1.0",
    "beautifulsoup4>=4.13.4",
    "lxml>=5.4.0",
]
